Drop f38 from image matrix
Ultimately we expect these images to be produced by the OS build
lifecycle, not via our CI.  That's in flight now.  Our primary
targets will anyways be eln and c9s; released Fedora actually
moves more slowly in many cases and that's not what we need
or are interested in here.