Unpin builder image
Let's assume for now that our own builder image isn't going
to break us.  This is needed to unblock getting `git` in the
image which we need for other things.

At some point yes we can investigate renovate for this potentially
but I'd rather say that we have decent CI around the builder image.