docs: Use reqpart in example.ks
The partitioning defined in the example kickstart file suggests that the
installer supports hybrid boot. That's misleading and not true. Let's use
the `reqpart` kickstart command to automatically create partitions required
by the detected platform instead of creating all of them for all platforms.

Note: The `reqpart` command doesn't work with `bootloader --location=none` or
`bootloader --disabled`, so this commit depends on the installer's support
for bootupd: https://github.com/rhinstaller/anaconda/pull/5298/