initramfs: Add virtiofs
Today it's built as a module for Fedora derivatives, but
it's a relatively small kmod.

In particular I want to be able to use a virtiofs root as
a bootstrap mechanism to generate disk images in a situation
where we can't do nested containers, but we do have `/dev/kvm` -
which is the setup we have in e.g. OpenShift Prow on `build02`
with nested virt.