system-configuration: Add cloud-utils-growpart
This is a relatively small shell script, and was already split out
from cloud-init to support use cases like ours.

systemd-repart is a lot nicer but it isn't yet designed
for our use case (root is not Discoverable Partition).

Also systemd-repart doesn't handle LVM which this script
tries to support.

(IMO the real solution to stuff like this is either Stratis
 or equivalent)