c9s: Just use production compose
I noticed we weren't getting updates and I think the
core problem here is renovate was only triggering on
the devel compose.  But we just "floated" to whatever
was found in the production compose as a base.

Also we had a random copr.

Drop that stuff, and just pin on the production compose.